calc: fix header context menu on touch devices
Previously we did not register or trigger the context menu when
long-pressing on the headers in calc on mobile devices. This commit
registers and activates the context menu when we detect a longpress

To avoid harming devices that are both touchscreen and mouse we do not
unregister the context menu after activating it (as on mouse-devices we
register/unregister the context menu based on the position of the
mouse). I believe this to never actually cause an issue however it's
plausible that triggering the context menu again on an area that doesn't
have a context menu could cause a context menu for a header to show. I
still think this is significantly better than the status-quo

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ib5e9cbdb3d7e10ed58396e8559ab2230f2b76c4c