Add rules about async/await to eslint
This commit adds 3 typescript-eslint rules about promises. For
no-floating-promises and no-misused-promises the rules are there to
avoid us forgetting await, which can lead to numerous issues such as
race conditions.

For require-await, the rule is there to avoid us marking functions as
async when they do not need to be. I've made it a warning to allow us to
ignore it in development, for example if we want to comment some async
code from an async function, but overall we shouldn't merge code that
contains unnecessarily-async functions.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I0ac42c760085e0f3dff77aaccba4742e0c2dd8ad